requirements.txt
fastapi>=0.115.12,<1.0.0
google-api-python-client>=2.169.0,<3.0.0
google-auth>=2.40.1,<3.0.0
google-auth-oauthlib>=1.2.2,<2.0.0
Pillow>=11.1.0,<12.0.0
python-dotenv>=1.0.1,<2.0.0
selenium>=4.29.0,<5.0.0
uvicorn>=0.34.2,<1.0.0
